commit : 최신화
